Script to refresh leetcode points for a class
At the moment, Codetrack API is not running a cron job to pull leetcode
points and this call is also not being made from the front end. For this
reason, if you need accurate point totals for the present time you will
need to manually hit the back end endpoint that does this for each learner.
This requires adding a `leetcode_id` field and value to each learner in
your roster. Then you can run `python codetrack_refresh_leetcode.py`